Fix: user int 64

+++ glaphyra/internal/app/users/dto/dto.go
@@ -22,7 +22,7 @@ const (
 )
 
 type User struct {
-	TgID             int32
+	TgID             int64
 	Username         string
 	Type             UserType
 	Style            string
